Ensure the middleware never breaks the chain of events
The middleware should only inject Adobe's Marketing Cloud Visitor ID on Screen and Track events
because they are the only ones that will be forwarded to Adobe Analytics server side integration.
All the other current or future types of events must be ignored and passed to the next middleware,
never breaking the chain of events.